+++ src/components/Button/Button.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { Props } from "../interfaces";
+import { Props } from "../../interfaces";
 const Button: React.FC<Props.ButtonProps> = ({ handleSave }) => {
   return (
     <div
