+++ src/components/index.ts
@@ -4,3 +4,5 @@ export { default as Button } from "./Button/Button";
 export { default as TextArea } from "./TextArea/TextArea";
 export { default as Table } from "./Table/Table";
 export { default as DropZone } from "./DropZone/DropZone";
+export { default as DropDown } from "./DropDown/DropDown";
+export { default as CheckBox } from "./CheckBox/Checkbox";
